Few fixes WorkLoadBalancer and Adding testcase for CapacityPlacement.

+++ src/main/java/com/vmware/borathon/Capacity.java
@@ -17,6 +17,6 @@ public class Capacity {
     }
 
     public double getCpuMemoryRatio(){
-        return (this.cpuMillicore/this.memoryMB);
+        return ((double)this.cpuMillicore/this.memoryMB);
     }
 }
